Tried to average age
I left a couple of comments. The ages need to be assigned at the same
time as the flux so that everything is the same length. Also, we only
need one age per spectrum. Right now we are pulling the entire array of
ages into each supernova. That's super bad.