Code sort of works
It successfully splits and plots by velocity and redshift...but there's
some sort of offset between them now. I think it has to do with the way
things were truncated and addressed. If we start with a complete
interpolated data set that has every bit of info instead of pulling from
three or four different tables, we can avoid that, I think.